requirements.txt
matplotlib>=2.1.2
numpy>=1.14.0
pandas>=0.22.0
tensorflow-gpu>=1.4.0
tqdm>=4.19.4
